+++ hw-5/src/main/java/ru/dankoy/hw5/core/commands/AuthorCommand.java
@@ -6,7 +6,7 @@ import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.shell.standard.ShellOption;
 import ru.dankoy.hw5.core.domain.Author;
-import ru.dankoy.hw5.core.service.ObjectMapperService;
+import ru.dankoy.hw5.core.service.objectmapper.ObjectMapperService;
 import ru.dankoy.hw5.core.service.author.AuthorService;
 
 @RequiredArgsConstructor
