--- hw-5/src/main/java/ru/dankoy/hw5/core/exceptions/MappingException.java
@@ -1,9 +1,0 @@
-package ru.dankoy.hw5.core.exceptions;
-
-public class MappingException extends Hw5RootException {
-
-  public MappingException(String message, Exception e) {
-    super(message, e);
-  }
-
-}
