+++ hw-5/src/main/java/ru/dankoy/hw5/core/exceptions/ObjectMapperException.java
@@ -1,0 +1,9 @@
+package ru.dankoy.hw5.core.exceptions;
+
+public class ObjectMapperException extends Hw5RootException {
+
+  public ObjectMapperException(String message, Exception e) {
+    super(message, e);
+  }
+
+}
